lisa.stats: Allow using group as rows in plot_stats() matrix
Using a new row for each subgroup (e.g. benchmark) can lead to explosion
of number of rows when there is a very large number of subgroup. If the
values of all subgroups happen to be of the same order of magnitude, it
can be interesting to use a row for each group, and have all subgroups
on the same bargraph.