lisa.datautils: Add df_delta() function
The df_delta function merges two dataframes of events
that are intended to be consecutive, e.g. the first 'pre'
event is emitted as the entry of a function, and a'post'
event is emitted at the exit.

The function keeps all the columns of the 'pre' and 'post'
dataframe. It adds an additional 'delta' column containing
the duration between the emission of a 'pre' event and its
consecutive 'post' event.

Signed-off-by: Pierre Gondois <Pierre.Gondois@arm.com>