lisa.datautils: add nullable parameter to series_convert()
series_convert() convert a pandas.Series to a given datatype.
The current behaviour of series_convert() is to convert the
serie to the nullable version of the datatype only if the
non-nullable version conversion failed.

This patch adds a new "nullable" parameter allowing to
force/prevent a conversion to a nullable datatype. Letting this
parameter unset keeps the current function's behaviour.

Reference:
https://pandas.pydata.org/pandas-docs/version/1.0.0/user_guide/integer_na.html

Signed-off-by: Pierre Gondois <Pierre.Gondois@arm.com>