lisa.trace: Use trace-cmd raw report for unknown events
BREAKING CHANGE

Events without a parser already in LISA will now default to parsing the
raw output of trace-cmd, since the human readable one will either be
identical or impossible to parse without a handcrafted parser.

In order to get the old behavior, a custom trace parser needs to be used
(or more likely, use a trace parser with custom event parser). See
Trace(parser=...) parameter documentation for how to do that.