lisa.trace: Fix default field conversions
Do not stop at the first conversion that seems to work, since it can end
up not giving the best dtype. For example, a column full of pointers
with MSB set will be converted to "object" dtype with
series_convert(series, 'int64'), which is not the best conversion we can
do. Instead, also try converting to "uint64" as well. If no conversion
fully succeeded, take the first conversion that did not raise.