lisa.wlgen.rta: Introduce new wlgen rt-app API
BREAKING CHANGE

Implement legacy classes as shims around the new classes.
The only breaking change is loops=-1 not being supported anymore in
RTATask, since rt-app provides no way of correctly composing workloads
with infinite loops (for now).

Also, all the new classes are designed to be immutable, which means that
inplace operations (e.g. +=) will return a new instance instead of
modifying the existing one.

Also split the responsibilities of RTAWorkload:
* running the workload => RTAWorkload
* creating the JSON configuration => RTAConf

This allows creating a configuration file without having to connect to a
target if needed.