tools: add '--native-build' option to lisa-build-asset
The '--native-build' option allows to natively compile an asset
instead of cross-compiling it. It uses alpine-chroot-install [1],
which installs Alpine Linux into a chroot and uses qemu to emulate
the desired architecture.

Only a few asset/architecture couples are currently available, so
the option only natively compiles an asset if possible.

As chroot requires root privileges, the '--native-build' option
also requires root privileges.

[1]: https://github.com/alpinelinux/alpine-chroot-install

Signed-off-by: Pierre Gondois <Pierre.Gondois@arm.com>
Signed-off-by: Douglas Raillard <Douglas.Raillard@arm.com>