lisa.tests.staging.utilclamp: ignore preemption in task activation
The test is skipping the first task activation, as the utilclamp values are
set afterward. However, if this first activation is preempted, the
preemption return would be seen as a new activation while it doesn't go
through CPU selection, leading to a test failure.

Ignoring preemption fixes this issue and keeps the test sane.