Vagrantfile: Use overlayfs for external/
Use an overlayfs for external/ to avoid hitting a virtualbox bug that
prevents pip from installing packages from the folder synced with the
host.

This means that modifications to external/ done from within the VM will
not be visible on the host. Changes made in the host may or may not be
visible inside the VM (there is no promise made by overlayfs). If
changes need to be visible, the VM should be rebooted first with
"vagrant reload"