lisa.wlgen.rta: List required kconfigs for taskgroups, uclamp
We currently try to reset uclamp values even when the kernel doesn't have
uclamp support compiled-in. This results in:

  [rt-app] <crit> [3] sched_setattr returned -1
  sched_setattr: failed to set uclamp value(s): Unknown error -1

(after some digging, the actual error is -95, aka -EOPNOTSUPP).

Check the kconfig and don't attempt to set default uclamp values if uclamp
isn't supported. Do the same thing for taskgroups.

Note that the warning is a bit spammy and might be better implemented with
a once-per-workload ratelimit.