lisa.energy_meter: Deprecate all classes
Devlib instruments should be used, and if there is a need to manage
configuration files on LISA side, we can add wrapping class like we did
for FtraceCollector and DmesgCollector.

The current state of the code would require a deep clean, but testing is
an issue because it requires specific hardware, so best done when it is
actually needed for a task.