lisa.tests.staging.utilclamp: Fix trace df windowing
Make sure we get a boundary at the beginning and end of the trace. We
can sometimes miss some phases if there was one continuous activation
starting from a given point, meaning there will not be any more context
switches until the end of the test. This can happen if cpufreq failed to
switch frequency, which therefore stayed too low and lead to
continuously missed activation.