lisa.analysis.frequency: Remove plot_peripheral_clock()
Remove as it is untestable without traces that have the right events.
Since it's a complex function, it would be impossible to port it to
holoviews without testing.

Please open an issue at https://github.com/ARM-software/lisa/issues if
you need it back, along with a trace.dat allowing the port to be done.