doc: lisa.wa.WATraceCollector: Fix example error
The example given for lisa.wa.WATraceCollector currently passes an
int to trace.analysis.idle.df_cluster_idle_state_residency(), which
expects a list(int). Because of that, attempts to run the example fail.

Wrapping the cpu id in a list during the function call solves the issue
without changing what the example is supposed to do.