lisa.analysis.tasks: Fix df_task_total_residency() for renaming tasks
Renamed PIDs currently lead to TaskIDs (or more):
    1. (PID, first comm)
    2. (PID, renamed comm)

Since task #1 finishes on an TASK_ACTIVE state, it will currently be
artifically extended until the end of the window, leading to an
artifically large runtime.

Fix that by introducing a new synthetic TaskState.TASK_RENAMED, that is
inserted on the "task_rename" event. This ensures we have an event that
is not TASK_ACTIVE at the end of the execution of tasks under their old
name. Note that this event is only introduced if the tasks are filtered
by comm, since it is only relevant in this scenario.