lisa.stats: Fix redundant tag columns handling
When two columns are redundant and only one of them is part of the
ref_group, this will lead to a situation where sub groups cannot be
compared their reference, since no subgroup of the refgroup matches.

Fix that by automatically extending ref_group with the value of the
redundant column.