lisa._di: Introduce DependencyInjector
FEATURE

Allow injecting dependencies into a module, which will create a fresh
instance of the module. This is mainly intended for parametrizing a
large amount of code with a number of classes, that can then be
user-provided at runtime.

In simpler case, a single class designed for subclassing to replace
parts is usually the way to go, but for larger amount of codes where
multiple classes are involved, it is actually easier to just code
against a set of abstract base classes and let the user fill the holes.