lisa.target: Enable root if username is not set
Class 'AndroidTarget' in devlib has removed the method
AndroidTarget::adb_root, thus it fails to execute 'adb root' for the
connection, and it complaints warning for "adb root" failure.

This leads to 'lisa-test' command execution with insufficient permission
to retrieve dmesg, and copy trace files.  To resolve this issue, one
option is to add the config 'username: root' into file target_conf.yml.

If the user doesn't set any username in the target config file, this
patch will automatically assign 'conn_settings['adb_as_root']' to 1, so
it can allow the adb connection to have root permission and dismiss the
testing failure.

Signed-off-by: Leo Yan <leo.yan@linaro.org>