lisa._unshare: Add lisa._unshare module
FEATURE

This module (private for now) provides an ensure_root() decorator, that
can be used to make sure a given Python function always run as root
user.

When called by a regular user, it will setup a new process with a
separated user and mount namespace and execute the function in it,
providing some sort of light container.