lisa.tests.scheduler.eas_behaviour: Use a custom RTAPhase Buffer
FIX

For the CFS policy, util_est implements a EWMA variable for the ramp-down.
It holds the utilization value and prevents convergence to a value matching the
duty cycle set in the rtapp profile. Using a custom RTAPhase with few tiny
activations make sure this value is reset and won't interfere with the
testing.