lisa.target: Remove adb rooting attempt
FIX

Remove attempt to call target.adb_root() when username is None. Devlib
API has changed and this is now a connection method. Calling that method
on the connection is not a fix as connections are created and destroyed
per thread, so applying it for the main thread is only kicking the can
down the road.

Since connecting as root can be achieved by simply using username='root'
and feeding that information into the connection settings, remove the
automatic attempt.