sched_tp: Handle a couple of compilation errors
sched_cpu_capacity depends on a tracepoint that was added on 5.10, so
protect it with the right kernel version.

Also the tracepoint makes a call to arch_scale_freq_capacity() which
doesn't compile on x86 as it points to non-exported function. Make the
trace event conditional on CONFIG_ARM64 which we know is supported and
produces intended output.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>