bisector: create new ArtifactoryService
ArtifactoryService is similar to ArtifactorialService, with the
exception that it requires a token (API key) for both upload and
download, which it provides through a header included in the request
sent to the server.

The upload folder and token will be retrived through
ARTIFACTORY_FOLDER and ARTIFACTORY_TOKEN env vars.

Artifactory also allows to specify properties for the uploaded
files, which can be provided as part of ARTIFACTORY_FOLDER.

Signed-off-by: Ionela Voinescu <ionela.voinescu@arm.com>