lisa.wlgen.rta: Add RTAPhaseBase.get_rtapp_repr(no_force_default_keys=None)
FEATURE

Allow not adding default value for some specified JSON keys. This is
useful for the "taskgroup" key as rt-app will error and bail if no
cgroup filesystem is mounted. Therefore, we cannot expect the default to
actually work everywhere and we must have a way of not adding it.