lisa.tests.scheduler.eas_behaviour: Frequency based duty cycle
FEATURE

Previously, SmallTask's would be described with a duty cycle of 50% of the
smallest CPU in the system. The util value might translate into a target
frequency too close from a real frequency. With PLET hazard, this could
lead or not to raising the selected frequency and reshuffle task placement
by EAS without LISA being aware of.

Instead of the hardcoded 50%, use a duty cycle that will translate into a
frequency right between two CPU frequencies. This gives a margin for PELT
fluctuation and mitigate the risk of unattended frequency selection.