lisa.monad: Allow composition of separate state monads
BREAKING CHANGE
FEATURE

Allow decorating a given coroutine function with more than one state
monad, in order to get the abilities brought by more than one API.

Also remove the automatic awaiting of monadic arguments to user-defined
coroutine functions. The reason is that it was a misfeature that worked
for the lisa.fuzz.Gen monad, but is actually quite problematic for other
state monads as the user looses control over the ordering of side
effects. This can unfortunately be extremely important for some monads.

Existing uses of the StateMonad should not be impacted beyond that.