lisa.wlgen.rta: Add async API
FEATURE

Add an async API to build wlgen rta tasks. The new API is based on the
previous one and its sole purpose is to provide a more intuitive front
end by somewhat using Python as a DSL.

The new API is not more or less powerful than the old one, and both can
be used together.