lisa._kmod: Fix PATH in Alpine chroot
FIX

Previously, PATH was not set/reset to any specific value when running a
command in the Alpine chroot. This lead to PATH being inherited from the
user, even though it did not specifically make sense inside the chroot.

Fix that by running "source /etc/profile" to provide a sane default for
basic env var.