tools/kmodules/: Make a symlink per file
FIX

Instead of a blanket symlink to lisa/_assets/kmodules/, create a
specific symlink per file. This avoids sharing the actual folder with
lisa/_assets/kmodules.

This separate is critical to ensuring the build of kmodules from
lisa._kmod is reliable, since files mtimes are lost when building from
Python, and out-of-date build artifacts can therefore sometimes be
reused.