doc/setup.rst: Fix manual kmod setup instruction
FIX

Do not refer to lisa/_assets/ anymore, refer to tools/kmodules/. This is
critical in ensuring lisa/_assets/ stays read-only with a pristine
content. Failure to do so leads to old artifacts (notably vmlinux.h)
being reused when they should not, as files mtime is lost when building
the module from Python code.

Since lisa/ folder is expected to be readonly anyway, redirect the user
to tools/kmodules where symlinks to specific source files are
established, rather than a blanket symlink for the whole folder.