lisa._kmod: Fix copy overlay backend
FIX

The copy overlay backend was not able to restore the deleted state of a
file when an upper was reused as a lower, like overlayfs is able to
using whiteout.

Fix that by only restoring the top-most lower layer when using the copy
backemd. This also has the advantage of speeding up the re-use of lowers
as only one tree needs to be copied.

After this change, lowers created with one backend cannot be mixed with
lowers created by another backend, so add the backend to the cache key
where appropriate.