lisa._assets.kmodules.sched_tp: Do not fail module load when no feature was manually selected
FEATURE

When the "features" kernel module parameter is not specified, do not
fail the module load in case some features fail to enable. This allows
enabling as many features as possible, which is likely what the user
expects.

If features were manually selected, we still fail the module load so the
user can rely on having the features they asked for or an obvious error.