lisa._assets.kmodules.sched_tp: Add MODULE_IMPORT_NS()
FIX

Some kernels require importing namespaces before using a symbol. An
example is some Android kernels that will refuse to load the module:

    [ 3357.743660] lisa: module uses symbol (filp_open) from namespace VFS_internal_I_am_really_a_filesystem_and_am_NOT_a_driver, but does not import it.
    [ 3357.756891] lisa: Unknown symbol filp_open (err -22)
    [ 3357.761935] lisa: module uses symbol (kernel_write) from namespace VFS_internal_I_am_really_a_filesystem_and_am_NOT_a_driver, but does not import it.
    [ 3357.775399] lisa: Unknown symbol kernel_write (err -22)
    [ 3357.780702] lisa: module uses symbol (kernel_read) from namespace VFS_internal_I_am_really_a_filesystem_and_am_NOT_a_driver, but does not import it.
    [ 3357.794067] lisa: Unknown symbol kernel_read (err -22)

Fix the issue by adding an MODULE_IMPORT_NS() for each namespace that
the kernel appears to be using, so that we are never at loss. This
allows compatibility with the widest range of kernels.