lisa: Use lisa.utils.destroyablecontextmanager
FIX

Equivalent to contextlib.contextmanager except that this pattern will
not run the cleanup code when context manager is destroyed without
calling __exit__() (explicitly or via "with" statement).

The usual pattern will run the cleanup code even if __exit__() is not
called, as destructing the generator will raise a GeneratorExit in it no
matter what:

    @contextlib.contextmanager
    def f():
	try:
	    yield
	finally:
	    cleanup()

This version will avoid that issue and will therefore behave like
context managers directly defined with of __enter__() and __exit__():

    @destroyablecontextmanager
    def f():
	try:
	    yield
	except ContextManagerExit:
	    cleanup()

This allows the final user of the context manager to skip cleanup if
they want to, which might be useful in some cases.