utilclamp.py: Don't consider migration margin in test_placement
By design, migration margin shouldn't be considered when comparing
uclamp with capacities as explained in this series:

	https://lore.kernel.org/lkml/20220804143609.515789-1-qais.yousef@arm.com/

It is considered a bug - that the above series fixes it.

Integration branch was failing with this series applied because of this
margin.

Fix the test to ignore the margin to reflect the actual uclamp design.

Older kernels without the series might fail now, which is the correct
result.

The series should trickle to stable kernels once merged in mainline
Linux. The failure should signal to any user that they need to backport
this series.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>