lisa.analysis: Add CPUs states DF generator
FEATURE

Add support to generated a DF reporting the consecutives timeframes in which a
CPU has been reported as IDLE (i.e. idle thread schediled in) or BUSY (i.e. non
running the idle thread).

These events are still useful to get a representation of the CPU status purely
based on the `sched_switch` event, e.g. when the cpu_idle events are not
available, either because not available or because not collected to optimized
the trace size.