lisa.analysis: Add CPUs states DF generator
FEATURE

Add support to generated a DF reporting the amount of time in a trace/view a CPU
has been BUSY or IDLE. Report also the amount of tracing time not covered by
events and thus "unaccounted" as BUSY/IDLE time. Finally, the overall CPU
utiliation for the a trace is also computed.