lisa.wa: Expose jobs, outputs in WAOutput & traces in WATraceCollector
FEATURE

This commit allows acessing a list of jobs in the WAOutput through
wa_output.jobs along with a LazyMapping of raw Trace objects in
WATraceCollector through wa_output['trace'].traces.

It also adds an 'outputs' property to WAOutput through which a dict of
discovered WAOutputs can be accessed.

Alongside that, a private _jobs property is created to unify job
discovery between the old multi-wa-output approach and the new
sigle-output approach.

It modifies the _needs_params method to ignore parameters that are
VAR_POSITIONAL or VAR_KEYWORD. This is so that the
trace collector can be accessed with ['trace'] instead of just through
get_collector() under this use case.