lisa.trace: Preserve time-range metadata in TxtTraceParserBase
FIX

time-range metadata used to be discarded unless asked for explicitly.
That likely comes from the time when we filtered per-event at the
trace-cmd level. Since we don't do that anymore since it prevents from
computing a globally unique timestamp for each event, there is no reason
to discard this information. That allows parsing a single event in a
single scan of the trace for most purposes.