lisa._kmod: Enable KBUILD_MODPOST_WARN=1
FIX

Module building fails with modpost error due to undefined symbols unless
KBUILD_MODPOST_WARN=1 is enabled, which turns the errors into warnings.
This is because modpost cannot know the list of symbols if
Module.symvers is missing, which is the case since we don't have a built
Image to work with.