lisa._assets.kmodules.sched_tp: parsec.h: Add private field to parse_buffer
This commit adds a 'private' field to parse_buffer in parsec.h. This
field can be used to pass down any data that could be relevant to the
parser's output but is not contained in the input itself.
It also adds a 'private' variable inside the SEQUENCE parser which
serves as a handle to that data.

Authored-by: Douglas Raillard <douglas.raillard@arm.com>
Signed-off-by: Kajetan Puchalski <kajetan.puchalski@arm.com>