lisa.trace: Fix FtraceCollector error when non-existant event is required
FIX

When an event is not available, FtraceCollector will try to build a
kernel module to try to provide it. If the module turns out to not
provide it, a None value will be propagated leading to an attribute
error.

Fix that by handling that case better and raise the expected
MissingTraceEventError.