lisa.trace: Fix task comm extraction
FIX

Trace.get_task_pid_names() returns the list of task names in
chronological order. We currently use the __comm and __pid columns of
some events as part of the data source. However, the __comm column is
the result of the lookup of __pid in a table stored in trace.dat, that
contains the latest comm known for a task. That effectively brings
information from the future and breaks the chronological order.

Instead, ignore those columns and focus on actual event fields.