exekall: Only consider callables in the packages under test
FIX

Root operators of expressions are only selected if they are defined in
a package to which belongs one of the module under test.

Extend that to all operators, as we might otherwise end up with another
library providing an operator for e.g. "int", leading to a combinatorial
explosion.