lisa._kmod: Rationalize the Alpine chroot arch selection
FIX

The alpine chroot build environment can use QEMU when a cross gcc
toolchain would be required, as Alpine does not ship a cross GCC
toolchain.

Refactor the logic to ensure that the binaries used are always in line
with the architecture of the chroot, and ensure QEMU is only used when
necessary.