lisa._kmod: Sanitize CC and make vars handling
FIX

Ensure selected compiler information is propagated everywhere it is
needed, such as when creating the Alpine chroot (to know what toolchain
to install and whether QEMU is necessary or not).

Also check that the user did not try to override the compiler at a let
stage after all the environment decisions have been taken.