lisa._kmod: Do not install gcc explicitly on clang Alpine chroot
FIX

When using clang, the toolchain from Alpine (or how it's used in KBuild)
still requires libgcc to link userspace tools successfully it appears.

Instead of installing gcc explicitly, install it indirectly via the
clang dependency so it will be naturally removed once the clang
toolchain does not require it anymore (hopefully).