lisa.trace: Add Trace(..., namespaces=("lisa", None)) parameter
FEATURE

Add support for event namespacing, which avoids name clashes while
allowing backward compatibility.

The namespaces to lookup are set on Trace() constructor and can be
overridden in Trace.df_event(namespaces=[...]). The full event name
looked up in a given namespace is <namespace>__<event>. Each namespace
specified will be tried in order until there is a match. None is a
special namespace that adds no prefix.

Note that setting the ftrace event subsystem is not enough to isolate
events, as they still live in a single "C source namespace" given that
the trace_* functions do not take the subsystem into account.