lisa.utils: Add !untrusted and !include-untrusted YAML tags to Serializable
FEATURE

Add an !untrusted YAML tag that allows including untrusted input in a
YAML document (e.g. using a templating system). The value must be a
string, which will be interpreted in isolation of the rest of the
document as a snippet of YAML.

Note that it's still on the user to ensure that the templating technique
used does not open the door to creating more than the string node, e.g.
by spanning multiple lines with less indentation.

!include-untrusted is similar to !include but for untrusted content.
Since the content is in a separate file, there is no risk of issue
coming from a specific string style.