lisa: Rename the module to "lisa" across all workflows
FEATURE

The lisa kernel module currently has varying names depending on how it
is built. It is called:

* "lisa" when built by the Python API.
* "sched_tp" when built manually.

The "sched_tp" name was preserved for backward compatibility, but that
is now hurting more than helping as Android ships its own version of
sched_tp, which corresponds to an ancient version of what used to exist
in LISA.

Consequently, the Android module and LISA module name clash when the
module is built manually for LISA. This also creates ambiguity when
documenting workflows, as the Android and LISA version of "sched_tp"
have diverged in requirements and features.

To remove any ambiguity, the LISA module is now renamed to "lisa" in all
workflows. Combined with the recent event rename (lisa__ prefix), both
Google's sched_tp and LISA's lisa module should be able to coexist
peacefully and independently of each other on the same system.