lisa._kmod: Use target.pull(as_root=True) when necessary
FIX

Pull root-owned files using as_root=True. This will ensure that pulling
works, even in cases where the server is not running as root (SSH
sometimes, adb almost always unless adb root was used).

Instead, the file will be copied to a temp location, its permission
changed (chmod runs with su or sudo so this can succeed even when the
daemon is not running as root) and then the temporary file pulled.