lisa.trace: Fix TraceEventCheckerBase.check_events() with namespaces
FIX

Ensure that we take into account the trace namespace when checking if an
event is in Trace.available_events. This object is not a regular set and
membership test will return True if the event is found in any of the
trace namespaces.

This is consistent with predicting the success or failure of
Trace.df_event().