lisa.conf: Deal with import errors more cleanly.
FIX

If a module declaring a configuration class fails to import, it might be
re-executed again (from another import statement). This currently leads
to an exception because two classes tried to register the same top-level
key, but they are actually the same class.

Instead of raising an exception, ignore that conflict and register the
latest type we encounter. This allows ensuring that the real import
error will surface, without extra noise from the conf system.