lisa/_assets/kmodules/lisa: Support in-tree builds
FEATURE

Modify the Makefile for the lisa module to distinguish
between in-tree and out-of-tree building. Adjust the
imports and definitions in the source files accordingly.

With this change, the module can now be copy-pasted into
a kernel tree and built with kbuild like any other kernel
module. The default, out-of-tree build method with Lisa
should work unchanged as before.

Signed-off-by: Kajetan Puchalski <kajetan.puchalski@arm.com>