tools/kmodules: Add helpers for in-tree module builds
FEATURE

Add patches that need to be applied to make in-tree Lisa
module building possible alongside a helper script.

Include patches to enable the Lisa module on product kernels
by re-exporting symbols forbidden by Google.

Update the documentation with instructions for in-tree builds
and instructions for the Lisa module pre-requisites.

Signed-off-by: Kajetan Puchalski <kajetan.puchalski@arm.com>