lisa.conf: Allow LevelKeyDesc to accept leaf values
FEATURE

LevelKeyDesc gained a value_path parameter that specifies the relative
path to a leaf key that will accept the value if a non-mapping value is
set for the level in input configuration.

This allows turning a leaf setting into a level in order to expand the
configuration options, while retaining backward compatibility on the
configuration file format. Note that this change will modify the layout
of the configuration so consuming code will need to be updated.