lisa._kmod: LISAFtraceDynamicKmod: Detect modules built in-tree
FEATURE

Override the install() method for LISAFtraceDynamicKmod to be able to
handle modules built in-tree and kept in the kernel modules directory.

The method will try to load the module using insmod while checking
whether the version the module was built with matches the Lisa checkout.
If the module version doesn't match, the build will fail with an error.
If the module can't be found in the modules directory, it will fall back
to the default way of building the module.

Signed-off-by: Kajetan Puchalski <kajetan.puchalski@arm.com>