tools/kmodules/lisa-in-tree: Adjust defaults and documentation
1. Add a usage example to the section in doc/setup.rst

Because we removed the default value for --module-kernel-path, it may
not be clear to the user what the argument should actually look like.
Add a usage example in the documentation to account for that.

2. Change the default remote from ssh to https

Defaulting to accessing the GitHub repo over ssh is prone to failing if
ssh is not set up on the machine. Https does not have that issue and so
is probably a more reasonable default.

3. Fix an omission in the kernel stub patch

The kernel stub patch still had `lisa-module-git` in the gitignore
where it should be `lisa-git`.

Signed-off-by: Kajetan Puchalski <kajetan.puchalski@arm.com>