lisa._kmod: Fix atomic headers sha1 fixup
FIX

CONFIG_IKHEADERS produces an archive of kernel headers that is
preprocessed to some extent to reduce its size. This processing
unfortunately breaks the sha1 checksum embedded in those headers by
changing the content of the file without updating the sha1.

Kbuild then checks that checksum and rejects that header. We are
therefore forced to fix the checksum embededd in the header.

So far, that fix was invasive in that it removed empty lines. That
unfortunately results in a dirty kernel tree when that recipe is applied
on a clean header. In that scenario (when CONFIG_IKHEADERS is disabled
but the recipe applied), we end up with a modified header, leading to a
"-dirty" uname, leading to a vermagic error when loading the module.

Fix that by making the checksum fixup less invasive.