lisa: Move test application code into its own package
BREAKING CHANGE

This is a breaking change from a setup point of view (separate package)
and when using "exekall run lisa" that now needs to be
"exekall run lisa lisa_tests" since the most class definitions are moved
to the new package.

However, from the API perspective this is not a breaking change, as a
later commit adds import shims (with a deprecation warning) in the old
locations.