tools/exekall: Handle namespace packages
FEATURE

If a valid path to a namespace package is provided, also import that
package as a Python name. This ensures that even if exekall is ran in
the repository of one part of that namespace package such that the path
and the Python name are equal, all the other parts will be found as
well.

Inference of the top-level package is not possible in general for
namespace packages as there is no __init__.py to be found in a parent
folder. However, when the module is being imported by name, we do know
what is the toplevel package so we can preserve that information when
infering the name of submodules.

Ensure namespace packages are added to sys.modules and as attribute on
the parent module.