tools/exekall: Add run --dependency option
FEATURE

Add a --dependency option that behaves the same as passing a module as
a positional argument, except that it will be ignored when computing the
list of allowed root operators.

This allows using a package to satisfy dependencies in expressions while
controlling what expressions are selected using the modules passed as
positional arguments.