tools/exekall: Restrict selected callables
FIX

Only select callables that are defined in the modules being scanned
(passed on the command line and any of their (indirect) children).

If a callable is defined in another module than the one being currently
scanned, it will be ignored. Since all the children of user-passed
packages are scanned, this has little effect, but allows more precise
selection when 2 packages or more are selected in case they depend on
each other.