lisa._kmod: Ensure all module sources are checksumed
FIX

Ensure that the entire set of sources are checksumed, including vmlinux
BTF blob and kallsyms if they are present. That is necessary as some
generated headers are derived from them, and what is available in the
module can change based on that so it has to be taken into account.