lisa._assets.kmodules.lisa: Use kallsyms symbol address
FIX

Instead of trying to know what symbol is exported (which is not doable
by just looking at kallsyms, since an STB_GLOBAL symbol does not imply
it has been exported), simply use the addresses from kallsyms to satisfy
the symbol references inside the module.

This has the advantage of somewhat circumventing the ABI export
restrictions that could be in place, but at the cost of relying on
accurate address from kallsyms. That is ensured by setting the following
files while we read /proc/kallsyms:

    /proc/sys/kernel/kptr_restrict set to 0
    /proc/sys/kernel/perf_event_paranoid set to -1