lisa.tests.staging.utilclamp: Avoid overlapping CPU capacities
FIX

Avoid choosing uclamp in CPU capacity ranges that are overlapping
between CPUs. Instead, ignore all the capacities on faster CPUs that are
lower than the max capa of any slower CPU.

Also ensure that the capacity bands considered exist inside a given CPU
and are not a band between the highest capacity of a CPU and the lowest
capacity of another one.