lisa._btf: Ensure the BTF types are dumped as deeply as possible when forward decls are used
FIX

Avoid accidentally stopping the recursive type printing when
encountering a forward declerence, since user code might be following a
pointer chain (a->b->c) where "b" is of an internal type, and suddenly
overnight becomes a forward decl, making "b->c" break.

Instead, resolve forward decl to the type they reference. This is
inaccurate since "struct foo;" is not necessarily a forward declaration
of "struct foo {...};" as they could both appear in totally unrelated
files.

However, the risk of that happening _in the set of types_ we care about
is quite low, and is lower than the risk of having forward decl popping
up randomly.