lisa._kmod: Cache compilation on cc --version
FIX

Caching of the compilation artifact was keyed on the compiler name so
far. This breaks if the user manipulates PATH and swaps different
binaries under the same name, e.g. multiple versions of clang all called
"clang".

Fix that by calling "cc --version" and caching based on the result. This
ensures we will not reuse artifacts created with a different toolchain.