lisa/utils: Fix wrong function signature in dispatch_kwargs()
Fixes bug in dispatch_kwargs where the object 'self' reference is passed
in kwargs causing 'self' to be dispatched twice: Positional argument and
kw argument.

The bug can be reproduced by creating an default RTA object, RTA().
Normally RTA objects are created through other functions, so the problem
only occurs when trying to create an RTA object from a provided rtapp
json file.

The bug is caused by inspect.signature() misbehaving for unbound
methods. This problem is already fixed in kwargs_dispatched(), so it
needs to be fixed in dispatch_kwargs() too.